pyproject.toml
[project]
name = "desafio-lacrei"
version = "0.1.0"
description = ""
authors = [
    {name = "Luis Miguel",email = "luismiguelsantossouza@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
pytest = "^8.4.1"
pytest-django = "^4.11.1"

